// src-ui/core/ipc/generated/debug/DebugActionCategory.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * 动作的分类。要 `Eq + Hash`（分组）
 */
export type DebugActionCategory = "message" | "group_info" | "group_admin" | "friend" | "file" | "request" | "account" | "face" | "stream" | "extension";